The monitor plugin template has been fixed and added to the default build tree, but excluded from the installation step. (#610)
