+++ app/mon/mon_plugins/plugin_template/src/metadata.json
@@ -1,7 +1,7 @@
 {
   "Version": "v0.1.0",
   "Author": "Author",
-  "Name": "Dummy Plugin",
+  "Name": "Template plugin",
   
   "Priority": 3,
