+++ app/mon/mon_plugins/plugin_template/src/plugin_widget.h
